chore: moved all chat related components into chat directory
